feat: Add script to create ACT Labs server environments
This commit adds a new script, create-actlabs-servers-envs.sh, which creates three ACT Labs server environments using the Azure containerapp command. Each environment is created with a unique name (actlabs-servers-env-01, actlabs-servers-env-02, actlabs-servers-env-03) and is associated with the "actlabs-servers" resource group and the "ACT-CSS-Readiness-NPRD" subscription. The environments are located in the "eastus" region and have their logs destination set to "none".